requirements.txt
discord.py==2.5.2
python-dotenv==1.0.1
discord.py==2.5.2
python-dotenv==1.0.1
pillow==11.1.0
psycopg2-binary